Rename and relocate tests
